+++ graphics/graphics.h
@@ -15,7 +15,7 @@ void save_picture(global_data *all_data);
 
 uint8_t getButton(int buttor_number, int position, int subpixel, int menuPosition);
 
-float getT(int k, int number_of_iterations);
+long double getT(int k, int number_of_iterations);
 
 uint8_t getR(float t);
 
@@ -29,13 +29,13 @@ void display_buffer(global_buffer * all_buffers, global_data * all_data);
 
 void erase_buffer(global_buffer * all_buffers, global_data * all_data);
 
-int iter_function(int number_of_iterations, float real, float imag, double c_real, double c_imag);
+int iter_function(int number_of_iterations, long double real, long double imag, long double c_real, long double c_imag);
 
 void compute_line(bool upperBool, global_buffer * all_buffers, global_data * all_data);
 
 void compute_column(bool leftBool, global_buffer * all_buffers, global_data * all_data);
 
-void compute_function(int cycle_start, int cycle_end, int local_width, int local_height, double c_real, double c_imag, int number_of_iterations, double min_real, double max_imag, double step_real, double step_imag);
+void compute_function(int cycle_start, int cycle_end, int local_width, int local_height, long double c_real, long double c_imag, int number_of_iterations, long double min_real, long double max_imag, long double step_real, long double step_imag);
 
 void cpu_compute(global_buffer * all_buffers, global_data * all_data);
